% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_history}
\alias{plot_history}
\title{Plot transaction and balance history.}
\usage{
plot_history(data)
}
\arguments{
\item{data}{A dataframe returned by read_statement*-functions.}
}
\value{
A ggplot-object.
}
\description{
Plot transaction and balance history.
}
