% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_partytotals}
\alias{plot_partytotals}
\title{Plot transactions totals by counterparty.}
\usage{
plot_partytotals(data)
}
\arguments{
\item{data}{A dataframe returned by read_statement*-functions.}
}
\value{
A ggplot-object.
}
\description{
Plot transactions totals by counterparty.
}
