% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statement.R
\name{read_statement}
\alias{read_statement}
\title{Read and clean a Nordea bank statement pdf-file.}
\usage{
read_statement(file)
}
\arguments{
\item{file}{A full path to a Finnish Nordea bank statement pdf-file.}
}
\value{
A dataset of balances and transactions of classes data.frame (base) and tbl_df (tibble).
}
\description{
\code{read_statement} reads and cleans up a Finnish Nordea bank statement
pdf-file and returns a tidy dataset of classes data.frame (base) and "tbl_df" (tibble).
}
