% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statements.R
\name{read_statements}
\alias{read_statements}
\title{Read and clean many Finnish Nordea bank statement pdf-files at once.}
\usage{
read_statements(directory)
}
\arguments{
\item{directory}{A full path to a directory of Finnish Nordea bank statement pdf-files.}
}
\value{
A dataset of transactions and balances of classes data.frame (base) and tbl_df (tibble).
}
\description{
This is just a convenient plural version of \code{read_statement}.
}
